mortgage_calculator/requirements.txt
flask
flask-cors
gunicorn
ngrok
python-dotenv
